+++ src/app/api/events/templates/route.ts
@@ -1,5 +1,6 @@
 import db from '@/dbconfig';
 import { eventService } from '@/features/events/services/event-service';
+import { eventTemplateService } from '@/features/events/services/event-template-service';
 import { loadSession } from '@/util/load-session';
 import { NextRequest, NextResponse } from 'next/server';
 
@@ -7,7 +8,7 @@ export async function GET(req: NextRequest) {
   try {
     const session = await loadSession();
     const searchParams = req.nextUrl.searchParams;
-    const templates = await eventService.repo.findTemplatesByAuthorId(
+    const templates = await eventTemplateService.repo.findTemplatesByAuthorId(
       session.user.id,
       searchParams.get('q'),
       db