+++ src/features/events/components/create-event-form/types-step.tsx
@@ -2,8 +2,12 @@ import { EventTypeSelector } from './inputs/event-type-selector';
 import { EventCategorySelector } from './inputs/event-category-selector';
 import { MobileEventCheckbox } from './inputs/mobile-event-checkbox';
 import { PositionAutoInput } from './inputs/position-auto-input';
+import { IsTemplateCheckbox } from './inputs/is-template-checkbox';
+import { useCreateEventForm } from '../../hooks/use-create-event-form';
+import { useCreateEventFormContext } from './create-event-form';
 
 export function TypesStep({ show }: { show: boolean }) {
+  const { template } = useCreateEventFormContext();
   return (
     <>
       <div
@@ -12,6 +16,7 @@ export function TypesStep({ show }: { show: boolean }) {
         <EventCategorySelector />
         <EventTypeSelector />
         <MobileEventCheckbox />
+        {!template && <IsTemplateCheckbox />}
         <PositionAutoInput />
       </div>
     </>